Separate vectorizer and add other features
