+++ medacy/nn/__init__.py
@@ -1,2 +1,3 @@
 """medacy.nn"""
 from .bilstm_crf import BiLstmCrf
+from .vectorizer import Vectorizer
